separate stories to files
